move convert_model to single device strategy
