+++ mmengine/_strategy/base.py
@@ -16,7 +16,6 @@ from mmengine.config import Config, ConfigDict
 from mmengine.dist import (broadcast, get_dist_info, infer_launcher,
                            is_distributed)
 from mmengine.logging import MMLogger
-from mmengine.model import revert_sync_batchnorm
 from mmengine.model.wrappers import is_model_wrapper
 from mmengine.optim import (BaseOptimWrapper, OptimWrapperDict,
                             _ParamScheduler, build_optim_wrapper)
@@ -301,23 +300,6 @@ class BaseStrategy(metaclass=ABCMeta):
         else:
             raise TypeError('model should be a nn.Module object or dict, '
                             f'but got {model}')
-
-    def convert_model(self, model: nn.Module) -> nn.Module:
-        """Convert layers of model.
-
-        convert all ``SyncBatchNorm`` (SyncBN) and
-        ``mmcv.ops.sync_bn.SyncBatchNorm`` (MMSyncBN) layers in the model to
-        ``BatchNormXd`` layers.
-
-        Args:
-            model (nn.Module): Model to convert.
-        """
-        self.logger.info(
-            'Distributed training is not used, all SyncBatchNorm (SyncBN) '
-            'layers in the model will be automatically reverted to '
-            'BatchNormXd layers if they are used.')
-        model = revert_sync_batchnorm(model)
-        return model
 
     def compile_model(
         self,
